this is my first changes

+++ AreaOfCircle.py
@@ -1,4 +1,4 @@
-radius=int(input("Enter the radius of circle="))
+radius=float(input("Enter the radius of circle="))
 
 Area=3.14*(radius*radius)
 print(Area)
